# Special-Case Ascending Value Types for System.Linq.OrderBy