@stephentoub 2020-01-06
@manofstick, thanks for the PRs.  I just returned from vacation.  I will review your two PRs hopefully this week.