@stephentoub 2020-06-08
@manofstick, I'm very sorry for the super-long delay on this one; sincere apologies.

In the end, I think we need to go ahead and close it.  I was keeping it open in hopes I could convince myself that we should take it, and it's great that it makes the common case with many elements faster, but it's pretty common with LINQ to process short enumerables and to do so repeatedly, and so a regression in a short list in favor of improvements in a long list isn't necessarily a good tradeoff.  But really the clincher for me is the long discussion that led to https://github.com/dotnet/runtime/pull/36643, where multiple folks in the community expressed strong concern at adding an O(N) amount of work even when doing so was part of removing an O(N log N) amount of work; this has a very similar feel to it, and I don't think the rewards associated with this PR are high enough to want to take such a risk again.

Thank you, again, for all of your efforts here.

cc: @eiriktsarpalis