@manofstick 2020-06-09
@stephentoub no worries :-) pretty hard to push anything through when there is a mature code base!

I mean I'm not 100% convinced that the arguments from #36643 really are relevant here (i.e. the predicate in that case was being used in places where it previously wasn't being used - i.e. slurped up into being called on all elements, rather than just the front elements of a sorted list - i.e. .net frameworks use has a deterministic state, i.e. the sorted first n elements until an event, but in PR it's just being used as an existing sort predicate where any assumptions of use by the client shouldn't exist (Otherwise *any* change to the sort algo would destroy clients, and I believe in the move to .net core as it moved from a vanilla quick sort to a slight modification, hence changing the potential order of the execution of the predicate, but that didn't cause issues)...

...but regardless, I can easily see from your perspective why this is a risk that you are unwilling to take.

Thanks for taking the time to have a look - who knows, I might even hit you with another PR soon :-P ! (there is one more from Cistern that I think could stand a chance!!)